moved the specific and general pages into student folder and edited the imports/routes to accomodate this refactoring
needed to be moved for easier lazy loading which loads only the needed module (loads only components needed for student and excludes those component for teacher)
