+++ src/app/app-routing.module.ts
@@ -10,12 +10,6 @@ import {
 
 //Application Imports
 import {
-  GenNewsComponent,
-  GenProfileComponent,
-  GenSelcourseComponent
-} from 'general/pages';
-
-import {
   LogInComponent
 } from 'log-in/log-in.component';
 
@@ -35,8 +29,8 @@ const routes: Routes = [
     pathMatch: 'full'
   },
   {
-    path: 'general',
-    loadChildren: './general/general.module#GeneralModule'
+    path: 'student/general',
+    loadChildren: './student/general/general.module#GeneralModule'
   },
   {
     path: 'log-in',
@@ -47,8 +41,8 @@ const routes: Routes = [
     component: SignUpComponent
   },
   {
-    path: 'specific',
-    loadChildren: './specific/specific.module#SpecificModule'
+    path: 'student/specific',
+    loadChildren: './student/specific/specific.module#SpecificModule'
   },
   {
     path: '**',
