+++ src/app/shared/services/page.service.ts
@@ -146,7 +146,7 @@ export class PageService {
 
 	public openSectionPage(sectionId: string) {
 		console.log(sectionId);
-		this.router.navigate(['/specific/specific-news', sectionId]);
+		this.router.navigate(['/student/specific/specific-news', sectionId]);
 	}
 
 	/**
