+++ src/app/student/general/gen-sidetab/gen-sidetab.component.css
