+++ src/app/student/general/gen-topnavbar/gen-topnavbar.component.html
@@ -6,13 +6,13 @@
   <div class="collapse navbar-collapse" id="collapsibleNavbar" [collapse]="isCollapsed">
     <ul class="nav navbar-nav">
       <li class="nav-item">
-        <a class="nav-link" (click)="isCollapsed = !isCollapsed" routerLink="/general/select-course">Courses</a>
+        <a class="nav-link" (click)="isCollapsed = !isCollapsed" routerLink="/student/general/select-course">Courses</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" (click)="isCollapsed = !isCollapsed" routerLink="/general/general-profile">Profile</a>
+        <a class="nav-link" (click)="isCollapsed = !isCollapsed" routerLink="/student/general/general-profile">Profile</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" (click)="isCollapsed = !isCollapsed" routerLink="/general/general-news">News</a>
+        <a class="nav-link" (click)="isCollapsed = !isCollapsed" routerLink="/student/general/general-news">News</a>
       </li>
     </ul>
     <ul class="navbar-nav ml-auto">
