+++ src/app/student/general/general.component.css
