+++ src/app/student/general/general.component.ts
