+++ src/app/student/general/pages/gen-news/gen-news.component.html
