+++ src/app/student/general/pages/gen-selcourse/gen-selcourse.component.ts
