+++ src/app/student/specific/pages/specific-quest-map/specific-quest-map.component.css
