+++ src/app/student/specific/specific-topnavbar/specific-topnavbar.component.html
@@ -6,24 +6,24 @@
   <div class="collapse navbar-collapse" id="collapsibleNavbar" [collapse]="isCollapsed">
     <ul class="nav navbar-nav">
       <li class="nav-item">
-        <a class="nav-link" [routerLink]="['/specific/specific-news', sectionId]">News</a>
+        <a class="nav-link" [routerLink]="['/student/specific/specific-news', sectionId]">News</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" [routerLink]="['/specific/specific-profile', sectionId]">Profile</a>
+        <a class="nav-link" [routerLink]="['/student/specific/specific-profile', sectionId]">Profile</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" [routerLink]="['/specific/specific-character', sectionId]">Character</a>
+        <a class="nav-link" [routerLink]="['/student/specific/specific-character', sectionId]">Character</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" [routerLink]="['/specific/specific-quest-map', sectionId]">Quest Map</a>
+        <a class="nav-link" [routerLink]="['/student/specific/specific-quest-map', sectionId]">Quest Map</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" [routerLink]="['/specific/specific-my-course', sectionId]">CMSC 128</a>
+        <a class="nav-link" [routerLink]="['/student/specific/specific-my-course', sectionId]">CMSC 128</a>
       </li>
     </ul>
     <ul class="navbar-nav ml-auto">
       <li class="nav-item">
-        <a class="nav-link" routerLink="/general/select-course" style="float:right"><i class="fa fa-arrow-circle-o-left"></i> Back to Courses</a>
+        <a class="nav-link" routerLink="/student/general/select-course" style="float:right"><i class="fa fa-arrow-circle-o-left"></i> Back to Courses</a>
       </li>
       <li cla ss="nav-item">
         <a class="nav-link" (click)="logOut()" style="float:right">Log Out</a>
